methods/active_minibatch/pyproject.toml
[project]
name = "gepa-active-minibatch"
version = "0.1.0"
description = "active_minibatch mutation: preferentially selects high-disagreement training examples"
requires-python = ">=3.12"
dependencies = ["gepa-mutations"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["active_minibatch"]
